requirements.txt
kafka-python==1.4.7
numpy==1.18.1
